Try to correct incomplete fix for kernel vs. userspace calling convention.  Seems like the current plan is to fix the calling convention in the libos, not libc.  Also, fix some cases where timeouts were being cast down to 32-bit integers

+++ Pal/src/host/Linux/db_mutex.c
@@ -56,7 +56,7 @@
 
 #define MUTEX_SPINLOCK_TIMES    100
 
-int _DkMutexLockTimeout (struct mutex_handle * m, int timeout)
+int _DkMutexLockTimeout (struct mutex_handle * m, uint64_t timeout)
 {
     int ret = 0;
 #ifdef DEBUG_MUTEX
